requirements.txt
paho-mqtt
discord.py
asyncio
python-dotenv